Begin porting oldstl List -> gc List layout
