+++ mycpp/gc_list_iter.h
@@ -1,0 +1,48 @@
+template <class T>
+class ListIter {
+ public:
+  explicit ListIter(List<T>* L) : L_(L), i_(0) {
+    // We need this because ListIter is directly on the stack, and L_ could be
+    // moved during iteration.
+    gHeap.PushRoot(reinterpret_cast<Obj**>(&L_));
+  }
+  ~ListIter() {
+    gHeap.PopRoot();
+  }
+  void Next() {
+    i_++;
+  }
+  bool Done() {
+    // "unsigned size_t was a mistake"
+    return i_ >= static_cast<int>(L_->len_);
+  }
+  T Value() {
+    return L_->slab_->items_[i_];
+  }
+
+ private:
+  List<T>* L_;
+  int i_;
+};
+
+// TODO: Does using pointers rather than indices make this more efficient?
+template <class T>
+class ReverseListIter {
+ public:
+  explicit ReverseListIter(List<T>* L) : L_(L), i_(L_->len_ - 1) {
+  }
+  void Next() {
+    i_--;
+  }
+  bool Done() {
+    return i_ < 0;
+  }
+  T Value() {
+    return L_->slab_->items_[i_];
+  }
+
+ private:
+  List<T>* L_;
+  int i_;
+};
+
